+++ lib/v2/screens/profile_screens/citizenship/interactor/usecases/get_referred_accounts_use_case.dart
@@ -1,0 +1,11 @@
+import 'package:async/async.dart';
+import 'package:seeds/v2/datasource/local/settings_storage.dart';
+import 'package:seeds/v2/datasource/remote/api/profile_repository.dart';
+
+class GetReferredAccountsUseCase {
+  final ProfileRepository _profileRepository = ProfileRepository();
+
+  Future<Result> run() {
+    return _profileRepository.getReferredAccounts('illumination');
+  }
+}
